Add ownership checks to memory loads and stores
In the extended compartmentalized memory model, load and store
operations (and their variants) take an additional parameter
representing the principal (component) that performs the operation. In
the simple model we are implementing, operations are allowed only if
the component owns the block, as given by the ownership information that
is already part of the memory.

The new ownership checks are added modularly to the relevant operations;
these are then propagated to related results. Many proofs need to be
modified, though typically the changes are straightforward. In general,
more case analysis is needed because conditions are more complex,
however the new sub-goals line up along predictable lines and the
main cases remain unchanged.

Perhaps the most interesting choice is that of compartments to add to
hypotheses and conclusions in modified lemmas: sometimes arbitrary,
sometimes limited to a distinguished component of interest, and other
times indifferent, though in practice bound by semantic constraints. In
some cases, ownership assumptions need to be added. These decisions will
need to be reviewed as the are used throughout the development, to make
their use as direct as possible.