Relate memory access modes and finish proofs
As it turns out, we actually have enough information to prove
interesting relations between the various permissions governing memory
accesses. Using these, we can complete the remaining goals in the
memory model and avoid pesky modifications to either the memory model
itself or to the definitions of some of its operations.

It remains to be seen how to keep visible interface changes to a minimum
and consider their impact on later proofs.