Extend compartments to common libraries
Most of the changes are or appear to be straightforward. A number of new
subgoals are generated; these mostly require us to prove new instances
of block ownership or relate compartments across operations. These goals
should all be straightforward.

Because compartments are sometimes hidden in some top-level
formulations, some tactics need to change to their existential variants
in order to work.

In addition to these general remarks, it should be possible to extend
hint databases to discharge some of the new goals, just as other
similar subproofs are found by simple proof search.