Start adding compartments to backend
The general strategy consists in adding compartments to all memory
operations, assigning one more more variables in as direct a manner as
possible, considering in particular the correspondence between
compartments in the assumptions and compartments in the conclusions.
While doing so, catalogue the unsolved goals and consider strategies
for discharging them efficiently. So far, nothing looks problematic or
difficult.