Extend compartmentalized memory to Stacking
The Stacking proof is perhaps the most interesting part of the backend,
in particular because the separation invariants on memories need to be
handled with some care. A conjunct stating memory ownership is added
together with the usual compartment parameter. The new proof obligation
is solved then instead of relegated, so that the unnecessary section
variables and hypotheses can be removed once the section is closed to
avoid interfering with external users. Nonetheless, some hypothesis
renumbering derived from the new conjunct seems difficult to avoid.

A small number of subgoals, in particular variants of old goals rather
than new goals, need a bit more work than usual to close all the gaps.