Memory compartments in Asm
The most interesting question is: where should we expose the existence
of compartments in the execution of Asm instructions? Unlike at other
levels the answer is not completely straightforward.

A possibility is to encapsulate these accesses inside [exec_instr],
essentially wrapping the original definition with checks to retrieve the
necessary (and readily available) information from the environment,
namely the PC component and its associated component. This choice
appears to be more self-contained, but certain results that require us
to establish, say, the successful execution of an instruction may
require additional assumptions (i.e., that the PC always contains a
pointer value and that its component lookup always succeeds).
Additionally, proofs on the wrapped version of [exec_instr] may have to
perform extra work to expose its underlying native form.

The other main possibility is to parameterize [exec_instr] on a
compartment and supply this information externally as needed. This
option requires a larger number of (minor) changes, but no additional
information. However, although the added compartment in [exec_instr]
needs to be added to all theorems using this definition, this parameter
is only of use for loads and stores, and unspecified in other cases. For
this reason, main helper lemmas, in particular [exec_straight_one], need
to be applied existentially, yet at the end their compartment variable
remains uninstantiated. Some local invariants need to be strengthened,
also straightforwardly, with compartment equality on memory operations.