Compartmentalized memory in frontend, first pass
This commits extends the first round of changes to the remaining
components of the frontend.

A point of interest is a result that points to the possibility of using
the [free_list] to free a list of blocks belonging to different
compartments, a possibility which was initially ruled out. At the
moment, adding the necessary constraints to respect this limitation
seems to work without issue.

The only point where significant adjustments could be expected shows up
in the proof of [make_memcpy_correct] in Chsmgenproof.v, where the
semantics of memcpy do not furnish sufficient information to connect
the owner of the source and destination blocks to the compartment of the
calling function. A fix will probably involve several semantic changes.

As usual, some additional ownership assumptions are added at various
locations. The most interesting of these is probably [tr_init_sound] in
Initializersproof.v.