Parameterize Cminor over compartments
As we need to connect Cminor expression evaluation with the compartment
in which it takes place, we add a compartment variable to the definition
of the Cminor language and propagate the changes. In most cases, there
is a clear target compartment to add to the evaluation, namely the
currently executing compartment of the function under execution.

Co-authored-by: Jérémy Thibault <jeremy.thibault@csp.mpg.de>