Begin to adapt backend pointers
Checkpoint on ValType refactoring for the new capability backend. Very
similar to CompCert standard values, refactoring the pointer part of the
definitions.

Proofs are very similar to those for standard pointers, sometimes with
added case analysis on the shape of the pointers (stack or heap). The
ValType specs line up better with this instance and all problematic
lemmas in the previous instance, except one, are proved right away.