Add signatures to call/return register functions
Below RTL, semantics move from an abstract, idealized view of
pseudo-registers to use of a concrete, finite set of machine registers.
In LTL and Linear, activation records are represented as collections of
locations (registers or stack slots). The semantics define auxiliary
functions to "automagically" save and restore registers according to the
ABI. Later, Mach concretizes activation frames even more.

The two modifications of interest ("caller-savification" and register
invalidation) are related, and easier to consider jointly.

At the upper levels, a clean solution can involve a modification of the
relevant auxiliary functions to perform register operations based not
only on the generic ABI, but also on the signature of the function being
called. Ultimately, this results in a sort of "variable" secure ABI
whose effects are function-specific -- and will need to be reflected at
lower levels as well.

As a first step, we add signatures to those helper functions. Simple
generalizations of existing results suffice to keep the proofs working
(for now).