Combined pass on LTL
Refine "secure calling convention" so that conversion of all registers
to caller-save and invalidation only takes place at cross-compartment
boundary crossings. This restores the original helpers that reflect the
calling convention and adds new ones for use at the call and return
thresholds. We will attempt to minimize changes to the structure of the
various language semantics starting with LTL. The modifications
introduce local case analyses on some of the call and return rules,
which will have a local effect on the proofs. At the language level, all
the information is available in the states being manipulated, with
several options to access the callee compartment to check the policy
(and with it the type of call) at call (or return) time.