Combined pass on Mach
As with the other languages, we try to make changes line up as much as
possible. The structure of stack frames undergoes significant
refinements, and with them register management.

The use of [comp_on_main] as a fallback on callee compartments on empty
stacks is particularly interesting, because we use it in the revised
definition of the semantics, i.e., the step function, and the addition
of the section variable modifies its standard signature. To avoid this,
at least until we explore the effects of the change and other relevant
uses, we alter the [callee_comp] fallback.