Checkpoint: sketch port of recomposition proof
The critical point is the definition of the mergeability relation and
the merge function. The existing proof relies on compartment memories,
which in turn make those definitions very simple. Here, because we are
intentionally staying as close to the original CompCert memory model as
possible, the same facilities no longer avail, and we need to do things
the "hard" way.