Experiment: universal register invalidation
Because the Allocation pass is a specialized type of static analyzer,
and because the call case makes assertions about sets equations not
constrained by (originally) callee-save registers, an extension of
this pass to a complex, context-sensitive calling convention is not
easily realized.

In order to select one calling convention (the standard ABI with
callee-save registers, which we would like to retain for
intra-compartment crossings) or the other (a secure calling convention
where every register is caller-save and non-argument registers are
moreover invalidated on compartment boundaries, which the proof needs
for the latter case), the static analyzer would need to determine when
a call or a return is changing compartments.

However, the static analyzer operates at the function level and is aware
of the function being analyzed (and its compartment). It does not know
which compartment(s) a call instruction may call into, because this
information is only available dynamically to the semantics based on the
current program state -- barring further restrictions or a more
sophisticated analysis.

For simplicity, we experiment with the extension that conservatively
applies register invalidation to all calls and returns. It remains to
evaluate the performance costs of this approach and the consequences on
the static analysis process itself.