Checkpoint: rework secure calling convention
Maintaining the standard RISC-V calling convention side by side with the
secure calling convention that classifies all machine registers as
caller-save and further invalidates them on cross-compartment calls and
returns is potentially a rather complex change. This is due to various
technical reasons.

The essence of the problem is being able to select which of the two
conventions should apply at any given call or return. This information
is not readily available from the call and return instructions
themselves, but is rather processed dynamically by the various
semantics.

The first pass where this distinction is of relevance, i.e. Allocation,
is also the point where this begins to create difficulties. The static
analyzer in this pass considers each function in isolation and cannot
determine statically whether a call or return is crossing a compartment
boundary or not. At the very least, additional annotations and source
restrictions would be needed to make these distinctions.

A further complication arises from the fact that the register allocator
itself is coded against the interfaces of the unified modules that
implement a single calling convention used throughout the entire
compiler. These interfaces are not rich enough to select one convention
or another based on boundary crossing information (which should be
available to begin with).

Consequently, the only way for the standard allocator to produce correct
results is to begin by applying the more restrictive, secure calling
convention to all calls. Conceptually, the modifications needed to
perform convention selection do not seem particularly complicated.

As noted, additional restrictions may be needed for the mixed calling
convention scenario to make sense. For example, even if we can determine
which calls to a given function cross compartment boundaries and which
do not, as long as both possibilities are allowed it would be difficult
for a single function to respond to both types of calls (e.g., saving
and restoring callee-save registers for internal calls, invalidating
non-return registers for external calls, etc.).

For the time being, we focus on getting the secure calling convention to
work and re-adapt the various definitions and proofs to work under this
new convention.