Propagate AST changes to compiler infrastructure
One change that had yet to be extended to the OCaml development is the
addition of compartments to all types of external functions.

The most pervasive changes are in C2C.ml, where helper functions are
extended with new compartment arguments, which are meant to be provided
by the function definitions and initializers that rely on those helpers.

Elsewhere the modifications are fairly simple. We should be mindful of
the role played by privileged (and formerly default) compartments.