Experiment with random graph generation
A graph effectively encodes a set of compartments and an edge
indicates that the two connected compartments import/export
functions from/to each other.

Also, for each node we sample a list of exported functions and
for each adjacent node we sample a random subset of the exported
functions which are then imported by the adjacent node.

For simplicity, neither the import nor the export list of any
compartment is ever empty.