Missing helpers and side hypotheses
An implicit assumption that we may need to make explicit is the
partition along the lines of a split between the context and the
programs. Related to these, a small number of lemmas relating global
environments and the initial memories derived from those are
needed; the current collection may not yet be optimal.